Surgically cut out the HTTP-related code to the `rest` sub-package and factored out common types and code into core, and happenings
