+++ main.go
@@ -6,12 +6,10 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"strconv"
 	"strings"
@@ -20,13 +18,12 @@ import (
 
 	hublleAgent "github.com/Jumpscale/hubble/agent"
 	hubbleAuth "github.com/Jumpscale/hubble/auth"
-	hublleProxy "github.com/Jumpscale/hubble/proxy"
 	"github.com/garyburd/redigo/redis"
-	"github.com/gin-gonic/gin"
-	influxdb "github.com/influxdb/influxdb/client"
 	"github.com/Jumpscale/agentcontroller2/core"
 	"github.com/Jumpscale/agentcontroller2/agentdata"
 	"github.com/Jumpscale/agentcontroller2/configs"
+	"github.com/Jumpscale/agentcontroller2/rest"
+	"github.com/Jumpscale/agentcontroller2/happenings"
 )
 
 const (
@@ -41,13 +38,6 @@ const (
 	cmdInternal               = "controller"
 )
 
-var influxDbTags = []string{"gid", "nid", "command", "domain", "name", "measurement"}
-
-//StatsRequest stats request
-type StatsRequest struct {
-	Timestamp int64           `json:"timestamp"`
-	Series    [][]interface{} `json:"series"`
-}
 
 // redis stuff
 func newPool(addr string, password string) *redis.Pool {
@@ -333,26 +323,14 @@ func cmdreader() {
 	}
 }
 
-var producers = make(map[string]chan *PollData)
+var producers = make(map[string]chan *core.PollData)
 var liveAgents = agentdata.NewAgentData()
 
 // var activeRoles map[string]int = make(map[string]int)
 // var activeGridRoles map[string]map[string]int = make(map[string]map[string]int)
 var producersLock sync.Mutex
 
-/*
-PollData Gets a chain for the caller to wait on, we return a chan chan string instead
-of chan string directly to make sure of the following:
-1- The redis pop loop will not try to pop jobs out of the queue until there is a caller waiting
-   for new commands
-2- Prevent multiple clients polling on a single gid:nid at the same time.
-*/
-type PollData struct {
-	Roles   []string
-	MsgChan chan string
-}
-
-func getProducerChan(gid string, nid string) chan<- *PollData {
+func getProducerChan(gid string, nid string) chan<- *core.PollData {
 	key := fmt.Sprintf("%s:%s", gid, nid)
 
 	producersLock.Lock()
@@ -364,7 +342,7 @@ func getProducerChan(gid string, nid string) chan<- *PollData {
 		//start routine for this agent.
 		log.Printf("Agent %s:%s active, starting agent routine\n", gid, nid)
 
-		producer = make(chan *PollData)
+		producer = make(chan *core.PollData)
 		producers[key] = producer
 		go func() {
 			//db := pool.Get()
@@ -382,7 +360,7 @@ func getProducerChan(gid string, nid string) chan<- *PollData {
 
 			for {
 				if !func() bool {
-					var data *PollData
+					var data *core.PollData
 
 					select {
 					case data = <-producer:
@@ -459,241 +437,10 @@ func getProducerChan(gid string, nid string) chan<- *PollData {
 	return producer
 }
 
-// REST stuff
-func cmd(c *gin.Context) {
-	gid := c.Param("gid")
-	nid := c.Param("nid")
 
-	query := c.Request.URL.Query()
-	roles := query["role"]
-	log.Printf("[+] gin: execute (gid: %s, nid: %s)\n", gid, nid)
 
-	// listen for http closing
-	notify := c.Writer.(http.CloseNotifier).CloseNotify()
 
-	timeout := 60 * time.Second
 
-	producer := getProducerChan(gid, nid)
-
-	data := &PollData{
-		Roles:   roles,
-		MsgChan: make(chan string),
-	}
-
-	select {
-	case producer <- data:
-	case <-time.After(timeout):
-		c.String(http.StatusOK, "")
-		return
-	}
-	//at this point we are sure this is the ONLY agent polling on /gid/nid/cmd
-
-	var payload string
-
-	select {
-	case payload = <-data.MsgChan:
-	case <-notify:
-	case <-time.After(timeout):
-	}
-
-	c.String(http.StatusOK, payload)
-}
-
-func logs(c *gin.Context) {
-	gid := c.Param("gid")
-	nid := c.Param("nid")
-
-	db := pool.Get()
-	defer db.Close()
-
-	log.Printf("[+] gin: log (gid: %s, nid: %s)\n", gid, nid)
-
-	// read body
-	content, err := ioutil.ReadAll(c.Request.Body)
-
-	if err != nil {
-		log.Println("[-] cannot read body:", err)
-		c.JSON(http.StatusInternalServerError, "error")
-		return
-	}
-
-	// push body to redis
-	id := fmt.Sprintf("%s:%s:log", gid, nid)
-	log.Printf("[+] message destination [%s]\n", id)
-
-	// push message to client queue
-	_, err = db.Do("RPUSH", id, content)
-
-	c.JSON(http.StatusOK, "ok")
-}
-
-func result(c *gin.Context) {
-	gid := c.Param("gid")
-	nid := c.Param("nid")
-
-	db := pool.Get()
-	defer db.Close()
-
-	log.Printf("[+] gin: result (gid: %s, nid: %s)\n", gid, nid)
-
-	// read body
-	content, err := ioutil.ReadAll(c.Request.Body)
-
-	if err != nil {
-		log.Println("[-] cannot read body:", err)
-		c.JSON(http.StatusInternalServerError, "body error")
-		return
-	}
-
-	// decode body
-	var payload core.CommandResult
-	err = json.Unmarshal(content, &payload)
-
-	if err != nil {
-		log.Println("[-] cannot read json:", err)
-		c.JSON(http.StatusInternalServerError, "json error")
-		return
-	}
-
-	log.Println("Jobresult:", payload.ID)
-
-	err = sendResult(&payload)
-	if err != nil {
-		log.Println("Failed queue results")
-		c.JSON(http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	c.JSON(http.StatusOK, "ok")
-}
-
-func stats(c *gin.Context) {
-	gid := c.Param("gid")
-	nid := c.Param("nid")
-
-	log.Printf("[+] gin: stats (gid: %s, nid: %s)\n", gid, nid)
-
-	// read body
-	content, err := ioutil.ReadAll(c.Request.Body)
-
-	if err != nil {
-		log.Println("[-] cannot read body:", err)
-		c.JSON(http.StatusInternalServerError, "body error")
-		return
-	}
-
-	// decode body
-	var payload []StatsRequest
-	err = json.Unmarshal(content, &payload)
-
-	if err != nil {
-		log.Println("[-] cannot read json:", err)
-		c.JSON(http.StatusInternalServerError, "json error")
-		return
-	}
-
-	u, err := url.Parse(fmt.Sprintf("http://%s", settings.Influxdb.Host))
-	if err != nil {
-		log.Println(err)
-		return
-	}
-
-	// building Influxdb requests
-	con, err := influxdb.NewClient(influxdb.Config{
-		Username: settings.Influxdb.User,
-		Password: settings.Influxdb.Password,
-		URL:      *u,
-	})
-
-	if err != nil {
-		log.Println(err)
-	}
-
-	points := make([]influxdb.Point, 0, 100)
-
-	for _, stats := range payload {
-		for i := 0; i < len(stats.Series); i++ {
-			var value float64
-			switch v := stats.Series[i][1].(type) {
-			case int:
-				value = float64(v)
-			case float32:
-				value = float64(v)
-			case float64:
-				value = v
-			default:
-				log.Println("Invalid influxdb value:", v)
-			}
-
-			key := stats.Series[i][0].(string)
-			//key is formated as gid.nid.cmd.domain.name.[measuerment] (6 parts)
-			//so we can split it and then fill the gags.
-			tags := make(map[string]string)
-			tagsValues := strings.SplitN(key, ".", 6)
-			for i, tagValue := range tagsValues {
-				tags[influxDbTags[i]] = tagValue
-			}
-
-			point := influxdb.Point{
-				Measurement: key,
-				Time:        time.Unix(stats.Timestamp, 0),
-				Tags:        tags,
-				Fields: map[string]interface{}{
-					"value": value,
-				},
-			}
-
-			points = append(points, point)
-		}
-	}
-
-	batchPoints := influxdb.BatchPoints{
-		Points:          points,
-		Database:        settings.Influxdb.Db,
-		RetentionPolicy: "default",
-	}
-
-	if _, err := con.Write(batchPoints); err != nil {
-		log.Println("INFLUXDB ERROR:", err)
-		return
-	}
-
-	c.JSON(http.StatusOK, "ok")
-}
-
-/*
-Gets hashed scripts from redis.
-*/
-func script(c *gin.Context) {
-
-	query := c.Request.URL.Query()
-	hashes, ok := query["hash"]
-	if !ok {
-		// that's an error. Hash is required.
-		c.String(http.StatusBadRequest, "Missing 'hash' param")
-		return
-	}
-
-	hash := hashes[0]
-
-	db := pool.Get()
-	defer db.Close()
-
-	payload, err := redis.String(db.Do("GET", hash))
-	if err != nil {
-		log.Println("Script get error:", err)
-		c.String(http.StatusNotFound, "Script with hash '%s' not found", hash)
-		return
-	}
-
-	log.Println(payload)
-
-	c.String(http.StatusOK, payload)
-}
-
-func handlHubbleProxy(context *gin.Context) {
-	hublleProxy.ProxyHandler(context.Writer, context.Request)
-}
 
 //StartSyncthingHubbleAgent start the builtin hubble agent required for Syncthing
 func StartSyncthingHubbleAgent(hubblePort int) {
@@ -746,7 +493,6 @@ func main() {
 
 	db.Close()
 
-	router := gin.Default()
 
 	go cmdreader()
 
@@ -759,30 +505,30 @@ func main() {
 
 	scheduler.Start()
 
-	eventHandler, err := NewEventsHandler(&settings.Events)
+	eventHandler, err := happenings.NewEventsHandler(&settings.Events, getProducerChan)
 	if err != nil {
 		log.Fatal(err)
 	}
+
+	restInterface := rest.NewRestInterface(
+		eventHandler,
+		getProducerChan,
+		pool,
+		sendResult,
+		&settings,
+	)
 
 	//start results processors
 	// processors := NewResultsProcessor(pool, resultsQueueMain)
 	// processors.Start()
 
 	hubbleAuth.Install(hubbleAuth.NewAcceptAllModule())
-	router.GET("/:gid/:nid/cmd", cmd)
-	router.POST("/:gid/:nid/log", logs)
-	router.POST("/:gid/:nid/result", result)
-	router.POST("/:gid/:nid/stats", stats)
-	router.POST("/:gid/:nid/event", eventHandler.Event)
-	router.GET("/:gid/:nid/hubble", handlHubbleProxy)
-	router.GET("/:gid/:nid/script", script)
-	// router.Static("/doc", "./doc")
 
 	var wg sync.WaitGroup
 	wg.Add(len(settings.Listen))
 	for _, httpBinding := range settings.Listen {
 		go func(httpBinding configs.HTTPBinding) {
-			server := &http.Server{Addr: httpBinding.Address, Handler: router}
+			server := &http.Server{Addr: httpBinding.Address, Handler: restInterface.Engine()}
 			if httpBinding.TLSEnabled() {
 				server.TLSConfig = &tls.Config{}
