+++ rest/cmd.go
@@ -1,0 +1,47 @@
+package rest
+import (
+"github.com/gin-gonic/gin"
+"log"
+"net/http"
+	"time"
+"github.com/Jumpscale/agentcontroller2/core"
+)
+
+func (r *RestInterface) cmd(c *gin.Context) {
+	gid := c.Param("gid")
+	nid := c.Param("nid")
+
+	query := c.Request.URL.Query()
+	roles := query["role"]
+	log.Printf("[+] gin: execute (gid: %s, nid: %s)\n", gid, nid)
+
+	// listen for http closing
+	notify := c.Writer.(http.CloseNotifier).CloseNotify()
+
+	timeout := 60 * time.Second
+
+	producer := r.producerChanFactory(gid, nid)
+
+	data := &core.PollData{
+		Roles:   roles,
+		MsgChan: make(chan string),
+	}
+
+	select {
+	case producer <- data:
+	case <-time.After(timeout):
+		c.String(http.StatusOK, "")
+		return
+	}
+	//at this point we are sure this is the ONLY agent polling on /gid/nid/cmd
+
+	var payload string
+
+	select {
+	case payload = <-data.MsgChan:
+	case <-notify:
+	case <-time.After(timeout):
+	}
+
+	c.String(http.StatusOK, payload)
+}
