+++ rest/stats.go
@@ -1,0 +1,115 @@
+package rest
+import (
+	"github.com/gin-gonic/gin"
+	"log"
+	"io/ioutil"
+	"net/http"
+	"encoding/json"
+	"net/url"
+	"fmt"
+	influxdb "github.com/influxdb/influxdb/client"
+	"strings"
+	"time"
+)
+
+//StatsRequest stats request
+type StatsRequest struct {
+	Timestamp int64           `json:"timestamp"`
+	Series    [][]interface{} `json:"series"`
+}
+
+var influxDbTags = []string{"gid", "nid", "command", "domain", "name", "measurement"}
+
+func (r *RestInterface) stats(c *gin.Context) {
+	gid := c.Param("gid")
+	nid := c.Param("nid")
+
+	log.Printf("[+] gin: stats (gid: %s, nid: %s)\n", gid, nid)
+
+	// read body
+	content, err := ioutil.ReadAll(c.Request.Body)
+
+	if err != nil {
+		log.Println("[-] cannot read body:", err)
+		c.JSON(http.StatusInternalServerError, "body error")
+		return
+	}
+
+	// decode body
+	var payload []StatsRequest
+	err = json.Unmarshal(content, &payload)
+
+	if err != nil {
+		log.Println("[-] cannot read json:", err)
+		c.JSON(http.StatusInternalServerError, "json error")
+		return
+	}
+
+	u, err := url.Parse(fmt.Sprintf("http://%s", r.settings.Influxdb.Host))
+	if err != nil {
+		log.Println(err)
+		return
+	}
+
+	// building Influxdb requests
+	con, err := influxdb.NewClient(influxdb.Config{
+		Username: r.settings.Influxdb.User,
+		Password: r.settings.Influxdb.Password,
+		URL:      *u,
+	})
+
+	if err != nil {
+		log.Println(err)
+	}
+
+	points := make([]influxdb.Point, 0, 100)
+
+	for _, stats := range payload {
+		for i := 0; i < len(stats.Series); i++ {
+			var value float64
+			switch v := stats.Series[i][1].(type) {
+			case int:
+				value = float64(v)
+			case float32:
+				value = float64(v)
+			case float64:
+				value = v
+			default:
+				log.Println("Invalid influxdb value:", v)
+			}
+
+			key := stats.Series[i][0].(string)
+			//key is formated as gid.nid.cmd.domain.name.[measuerment] (6 parts)
+			//so we can split it and then fill the gags.
+			tags := make(map[string]string)
+			tagsValues := strings.SplitN(key, ".", 6)
+			for i, tagValue := range tagsValues {
+				tags[influxDbTags[i]] = tagValue
+			}
+
+			point := influxdb.Point{
+				Measurement: key,
+				Time:        time.Unix(stats.Timestamp, 0),
+				Tags:        tags,
+				Fields: map[string]interface{}{
+					"value": value,
+				},
+			}
+
+			points = append(points, point)
+		}
+	}
+
+	batchPoints := influxdb.BatchPoints{
+		Points:          points,
+		Database:        r.settings.Influxdb.Db,
+		RetentionPolicy: "default",
+	}
+
+	if _, err := con.Write(batchPoints); err != nil {
+		log.Println("INFLUXDB ERROR:", err)
+		return
+	}
+
+	c.JSON(http.StatusOK, "ok")
+}
